user table crud and server refactor
