+++ src/app.ts
@@ -3,8 +3,8 @@ import { server } from '@/server/server';
 
 import '@/routes/index.route';
 import '@/routes/health.route';
-import '@/routes/task.route';
-import '@/routes/user.route';
+import '@/domains/tasks/task.route';
+import '@/domains/users/user.route';
 import '@/routes/protected.route';
 
 configureOpenAPI(server);
