+++ src/db/schema.ts
@@ -1,5 +1,6 @@
 import { z } from '@hono/zod-openapi';
 import { boolean, pgTable, serial, text, timestamp } from 'drizzle-orm/pg-core';
+import { sql } from 'drizzle-orm';
 import { createSchemaFactory } from 'drizzle-zod';
 
 export const tasks = pgTable('tasks', {
@@ -13,15 +14,16 @@ export const tasks = pgTable('tasks', {
 });
 
 export const users = pgTable('users', {
-  id: serial('id').primaryKey(),
-  name: text('name').notNull(),
-  email: text('email').notNull().unique(),
-  role: text('role').notNull().default('user'),
-  active: boolean('active').notNull().default(true),
+  id: text('id').primaryKey().default(sql`gen_random_uuid()`),
+  username: text('username').notNull(),
+  email: text('email').notNull(),
+  firstName: text('first_name'),
+  lastName: text('last_name'),
+  role: text('role').notNull(), // UUID stored as text
+  createdBy: text('created_by'), // UUID stored as text
   createdAt: timestamp('created_at').defaultNow(),
-  updatedAt: timestamp('updated_at')
-    .defaultNow()
-    .$onUpdate(() => new Date()),
+  updatedAt: timestamp('updated_at').defaultNow(),
+  isActive: boolean('is_active').notNull().default(true),
 });
 
 const { createInsertSchema, createSelectSchema } = createSchemaFactory({
@@ -47,18 +49,47 @@ export const patchTasksSchema = insertTasksSchema.partial();
 export const selectUsersSchema = createSelectSchema(users);
 
 export const insertUsersSchema = createInsertSchema(users, {
-  name: (str) => str.min(1).max(100),
+  username: (str) => str.min(1).max(100),
   email: (str) => str.email(),
-  role: (str) => str.min(1).max(50),
+  firstName: (str) => str.min(1).max(100).optional(),
+  lastName: (str) => str.min(1).max(100).optional(),
+  role: (str) => str.uuid(), // Role as UUID
 })
   .required({
-    name: true,
+    username: true,
     email: true,
+    role: true,
   })
   .omit({
     id: true,
     createdAt: true,
     updatedAt: true,
-  });
+    createdBy: true,
+  })
+  .transform((data) => ({
+    ...data,
+    // Convert empty strings to null for optional fields
+    firstName: data.firstName === '' ? null : data.firstName,
+    lastName: data.lastName === '' ? null : data.lastName,
+  }));
 
-export const patchUsersSchema = insertUsersSchema.partial();
+export const patchUsersSchema = createInsertSchema(users, {
+  username: (str) => str.min(1).max(100),
+  email: (str) => str.email(),
+  firstName: (str) => str.min(1).max(100).optional(),
+  lastName: (str) => str.min(1).max(100).optional(),
+  role: (str) => str.uuid(), // Role as UUID
+})
+  .omit({
+    id: true,
+    createdAt: true,
+    updatedAt: true,
+    createdBy: true,
+  })
+  .partial()
+  .transform((data) => ({
+    ...data,
+    // Convert empty strings to null for optional fields
+    firstName: data.firstName === '' ? null : data.firstName,
+    lastName: data.lastName === '' ? null : data.lastName,
+  }));
